feat: variant 분리
