+++ src/components/ui/Tag/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Tag';
+export * from './tag.variant';
